requirements.txt
twikit
openai
geopy
python-dotenv
requests
tweepy
